Change message processor Fix memory leaks Increase wsring level
git-svn-id: file:///.data/svn_convert/svn@2626 f7a67266-ba00-4dfb-90a2-3eec4ba07b7d

+++ branches/unlabeled-1.2.2/plugins/_core/phonedetails.cpp
@@ -46,7 +46,7 @@ PhoneDetails::PhoneDetails(QWidget *p, const QString &oldNumber)
         number = number.mid(0, pos);
     }
     number = trim(number);
-    initCombo(cmbCountry, countryCode, getCountries());
+    initCombo(cmbCountry, (unsigned short)countryCode, getCountries());
 
     QFontMetrics fm(font());
     unsigned wChar = fm.width("0");
