+++ branches/unlabeled-1.7.2/plugins/icq/icqconfig.cpp
@@ -76,9 +76,9 @@ void ICQConfig::apply()
         m_client->setPassword(edtPasswd->text());
     }
     m_client->setServer(edtServer->text().local8Bit());
-    m_client->setPort(atol(edtPort->text()));
-    m_client->setMinPort(atol(edtMinPort->text()));
-    m_client->setMaxPort(atol(edtMaxPort->text()));
+    m_client->setPort((unsigned short)atol(edtPort->text()));
+    m_client->setMinPort((unsigned short)atol(edtMinPort->text()));
+    m_client->setMaxPort((unsigned short)atol(edtMaxPort->text()));
     m_client->setSendFormat(cmbFormat->currentItem());
     m_client->setAutoUpdate(chkUpdate->isChecked());
     m_client->setAutoReplyUpdate(chkAutoReply->isChecked());
